+++ cherimoya_cli/defaults.py
@@ -114,7 +114,7 @@ default_evaluate_parameters = {
 
 
 default_attribute_parameters = {
-	'batch_size': 512,
+	'batch_size': 64,
 	'in_window': 2114,
 	'verbose': False,
 	'chroms': training_chroms + validation_chroms,
@@ -122,10 +122,15 @@ default_attribute_parameters = {
 	'sequences': None,
 	'loci': None,
 	'model': None,
-	'compile': True,
+	'compile': False,
 	'compile_mode': 'max-autotune',
+	'algorithm': 'deep_lift_shap',
 	'output': 'counts',
+	'group': 0,
 	'attr_window': 400,
+	'n_shuffles': 20,
+	'warning_threshold': 1e-3,
+	'random_state': 0,
 	'ohe_filename': 'attributions.ohe.npz',
 	'attr_filename': 'attributions.attr.npz',
 	'idx_filename': 'attributions.idx.npy',
@@ -274,10 +279,16 @@ default_pipeline_parameters = {
 
 	# Attribution parameters
 	'attribute_parameters': {
-		'batch_size': None,
+		'batch_size': 64,
+		'compile': False,
 		'chroms': training_chroms + validation_chroms,
+		'algorithm': 'deep_lift_shap',
 		'output': 'counts',
+		'group': 0,
 		'attr_window': None,
+		'n_shuffles': 20,
+		'warning_threshold': 1e-3,
+		'random_state': None,
 		'loci': None,
 		'dtype': None,
 		'device': None,